backup_20250726/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
numpy
sympy
scipy
matplotlib